+++ src/physics/impact/SeparateX.js
@@ -7,7 +7,7 @@
 /**
  * [description]
  *
- * @function Phaser.Physics.Impact.SeperateX
+ * @function Phaser.Physics.Impact.SeparateX
  * @since 3.0.0
  *
  * @param {Phaser.Physics.Impact.World} world - [description]
@@ -15,7 +15,7 @@
  * @param {Phaser.Physics.Impact.Body} right - [description]
  * @param {Phaser.Physics.Impact.Body} [weak] - [description]
  */
-var SeperateX = function (world, left, right, weak)
+var SeparateX = function (world, left, right, weak)
 {
     var nudge = left.pos.x + left.size.x - right.pos.x;
     
@@ -47,4 +47,4 @@ var SeperateX = function (world, left, right, weak)
     }
 };
 
-module.exports = SeperateX;
+module.exports = SeparateX;